capi: Add object sizes to input types
In order to ensure some level of forward & backward compatibility, this
change extends all our input types with a member holding the objects
size. Users need to set this member to the size of the type. The library
can then ignore unsupported fields (forward compatibility) or provide
reasonable defaults for fields not provided (backward compatibility).

Signed-off-by: Daniel Müller <deso@posteo.net>