capi: Ensure that input types are zero padded
Because blazesym-c aims to be forward compatible, it needs to accept
objects that are larger (i.e., include extensions) than what the library
currently defines. However, the library should not attempt to satisfy a
request that has options present that it does not understand.
To enforce this constraint, put in some checks that make sure that any
bytes past the last currently defined member (but capured by the _size
field) are set to zero.

Signed-off-by: Daniel Müller <deso@posteo.net>