proc macro -- NOT WORKING
This can be made to work for functions etc, but not for members in a
struct or variants in an enum: attribute macros are not available in
those positions and derive macros do not allow for replacement of the
type definition, only appending (they are "inert" as opposed to
"active"). Ugh.