Add rust-minidump based Breakpad parsing logic
In the future we would like to support the Breakpad format for use as a
symbolization source. The closest suitable crate providing this
functionality is breakpad-symbols [0]. However, it contains a slew of
other functionality that we have no use for, including a HTTP based symbol
retriever, a stack walker, as well as plumbing & dependencies for async
APIs. On top of that, their parsers error reporting is insufficient at
best.
With this change we pull in a snapshot of the core parser logic and
adjust it as per our needs. Right now that mostly means removing of
async bits, eliminating the range-map and various other dependencies,
and removing unnecessary bits from the parser (e.g., for stack unwinding
information). The snapshot is based on rust-minidump [1] at
commit a0689028d5b13232c82141672edc3f4d70817bd7.
This functionality is still unused. Subsequent changes will add a
resolver that relies on it.

[0] https://crates.io/crates/breakpad-symbols
[1] https://github.com/rust-minidump/rust-minidump

Signed-off-by: Daniel Müller <deso@posteo.net>