Add support for inlined function reporting to Breakpad
This change adds support for inlined function reporting to our Breakpad
logic. Note that for this format inlined functions are currently parsed
(and stored) unconditionally when the entire file is parsed. That is
different to other formats that try to do that in a more lazy fashion.
We could potentially refine the logic later on, but the effort does not
justify the gains at this point. We still honor the user's desire and
only look it up and report it if asked for, of course.

Signed-off-by: Daniel Müller <deso@posteo.net>