Introduce "Resolver" type?
The problem:
we mix container/virtual symbolization sources (APK, process, kernel)
with concrete one (Breakpad, ELF, Gsym, Ksym, ...). This is inflexible
API design and is causing friction in niche use cases. E.g., when
attempting to symbolize an address in an APK using a Breakpad belonging
to one of the contained ELF files (instead of ELF/DWARF symbols embedded
inside the ELF file). Or, using out-of-band Gsym debug information when
symbolizing addresses inside a process. In many ways, it is a
generalization of problems we envision having to solve once we support
split DWARF information.

First idea:
Split out a "Resolver" type (name TBD; conflicts with SymResolver
trait), that can be used to map an address on such a "symbolic" symbol
source to a concrete Resolver instance.
- problems: it's tricky to establish an interface between Symbolizer &
  Resolver and the caching story gets tricky
  - do we return an owned, heap allocated object?
    - that's not great for "high-throughput" use cases such as process
      symbolization

Second idea:
Configure a Symbolizer instance with functions for the container formats
(at construction time). E.g., a user could set
Builder::set_apk_resolver_fn() and install a function
fn(apk_path: &Path, member_path: &Path, member_data_bounds: Range<u64>)
-> Option<Rc<dyn SymResolver>>

This function could be used to a SymResolver instance that is to be used
for any addresses falling onto this APK member. The Symbolizer instance
could cache this resolver.

The implication of this approach would be that the entire SymResolver
trait and the relevant methods find_sym() & find_code_info() would need
to be made public. This has a slew of trickle down effects, because the
methods currently return crate private types (and for good reason).