Make DwarfResolver fall back to using ELF symbols
So far DWARF logic has just been treated as a sub-set (or special case?)
of ELF, with an opt-in being required to peek into DWARF symbols. That
is questionable on many fronts. For one, usage of only ELF symbols
should be the exception and not the norm: DWARF typically provides
everything that ELF offers and much more (e.g., source code location
information and inlined function support). There is a trade-off to be
made with respect to memory consumption and reporting speed, but
favoring ELF from the get-go without a hard reason why seems
questionable: symbolization using ELF is perhaps an order of magnitude
faster than with DWARF, which is not insignificant, but it pales
compared to Gsym performance (which beats ELF by >2 orders of
magnitude?). Hence, if speed really were the decisive factor then Gsym
would beat ELF hands down (it also supports source code location
reporting etc.).
Long story short, we with upcoming changes we are going to invert the
relationship between ElfResolver and DwarfResolver: DwarfResolver will
be able to fall back to using ELF symbols but an ElfResolver has no
knowledge of DWARF. This change marks a first step towards such a world,
adjusting the DwarfResolver type to fall back to using ELF symbols if
DWARF information was insufficient to serve the request.

Signed-off-by: Daniel Müller <deso@posteo.net>