Introduce Resolve trait
This change introduces the Resolve trait. This will become the main user
facing trait that client code may opt satisfy in order to hook into the
symbolization path. Alongside we have to introduce the AsSymbolize
trait, which facilitates conversion of a Resolve instance into a
Symbolize, while generic trait upcasting is not yet available in stable
Rust. We opt to switch the APK cache over from storing ElfResolver
objects to the more generic dyn Resolve, in anticipation of future
changes.

Signed-off-by: Daniel Müller <deso@posteo.net>