Add support for converting Box<dyn StdError> into Error
While generally it is only the library that creates its Error objects,
with upcoming changes we want to provide users with more control over in
advanced usage scenarios and in those cases it may be necessary to
convert arbitrary errors into blazeym's.
With this change we add a conversion from Box<dyn StdError> to
blazesym's Error type, that will offer the aforementioned flexibility to
users.

Signed-off-by: Daniel Müller <deso@posteo.net>