Make Kfunc to ResolvedSym conversion fallible
Adjust the functionality for converting from Kfunc to ResolvedSym to be
fallible and stop relying on the From trait for that purpose. We will
eventually end up including a bit more information in the conversion and
so let's instead roll with a dedicated conversion method on the Kfunc
type itself. We need fallibility because this new method acts as an
"abstract interface" for additional implementations coming in the future
and which will be fallible.

Signed-off-by: Daniel Müller <deso@posteo.net>