Pull in vmlinux.h functionality
With upcoming changes we are going to need to test functionality using
BPF. BPF is a royal pain to work with, because it requires various
system dependencies.
This change addresses one of these dependencies: in order to build BPF
programs we are going to need vmlinux.h available. We add infrastructure
for pulling in vmlinux.h from a centrally managed location for building
BPF test programs.

Signed-off-by: Daniel Müller <deso@posteo.net>