capi: Add support for capturing traces
The main Rust crate allows for configuration of the generic tracing
infrastructure to glean some insights into what is going on. So far the
C library didn't have any mechanism for tapping into that.
This change adds such infrastructure with the new blaze_trace()
function, which allows for registration of a callback function that is
invoked for each emitted trace line.

Closes: 587

Signed-off-by: Daniel Müller <deso@posteo.net>