Incorporate ELF symbols of unknown type in symbolization
The symbolize_kernel_system_vmlinux test flagged another case of symbols
that we may want to handle better: some ELF symbols have type NOTYPE,
which according to the specification means "The symbol's type is not
defined". So far we have outright excluded these symbols from
symbolization, but it seems better to include them, as they could refer
to either variables or functions (or, potentially anything else there
is...). Presumably they refer to unannotated definitions made directly
in assembly. An example is:
```
 63755: ffffffff81601120   117 NOTYPE  GLOBAL DEFAULT    1 entry_SYSENTER_compat
```

With this change we make sure to include these symbols when parsing
symbol tables and report them when asking for both functions and
variables.
Unfortunately, getting our file offset tests passing with this change
resulted in a hard-mode game of whack-a-mole. The conclusion we came to
was that file offset calculation for symbols in SHT_NOBITS sections
should return None, because when section contents are not backed by data
in the file, it can't really have a file offset. However, that breaks
file offsets for debug linked files, for which ELF sections are
typically patched to have the SHT_NOBITS type. For now we won't report
file offsets for these.

Signed-off-by: Daniel Müller <deso@posteo.net>