capi: Trace errors on unzeroed input structs
For forward compatibility, we expect input structs to be zeroed past
the corresponding type's objects size. It shouldn't be too hard to
recognize when this constraint is violated at the call site, but its
also not terribly hard for us to be a tad more verbose about this fact.
Add a few traces to aid with debugging.

Signed-off-by: Daniel Müller <deso@posteo.net>