Remove BuildIdReader trait
The BuildIdReader was a useful and clean abstraction when introduced.
However, with upcoming changes we want to read build IDs of ELF files
inside APKs, at which point the interface (which works on Path objects)
breaks down.
Remove the BuildIdReader trait and its various implementations, in favor
of a mere method for reading a build ID.