Assume system-wide vDSO image
When we implemented the vDSO functionality the assumption was that the
vDSO could potentially be different between different processes. The
belief was that it is provided by libc and with some legwork it is
possible to use a different libc per-process.
In reality, however, the vDSO is solely managed by the kernel. The
contract with libc is just that libc assumes a minimum kernel version
and available vDSO function based on that. Given that only a single
kernel can be in use on a system, we can simplify our vDSO symbolization
logic quite a bit and even relax the required permissions.

Signed-off-by: Daniel Müller <deso@posteo.net>