Manage FileCache entries more globally
The FileCache type is used to keep around data associated with a certain
"version" of a file, i.e., a given snapshot at a certain time, uniquely
identified by various stat(2) attributes. So far we managed all the
data associated with the different versions at the path level. That is,
each path had a map of <version, entry> associated with it. Doing so
basically meant that entries could only ever be associated with a single
path. That is a problem when it comes to symbolic links, because with a
symbolic link we may want to associate an entry with multiple paths.

With this change we move the entry map from the individual PathEntry
object into the top level FileCache, managing all of them in a "global"
manner. An additional positive side effect of this change is that the
not-insignificant memory size of an InsertMap is now no longer duplicate
for each path.

Signed-off-by: Daniel Müller <deso@posteo.net>