Overhaul symbolic link handling in FileCache
Rework the handling of symbolic links in our FileCache. So far we
effectively ignored them, always working with the target file, but still
duplicating entries because of our look up by name.
With this change we make sure to properly detect and follow symbolic
links and deduplicate equal target files to the same cache entry.

Closes: #1165

Signed-off-by: Daniel Müller <deso@posteo.net>