Store kernel module names in Kfunc type
Upcoming changes aim at improving our kernel symbolization logic for
kernel modules, so that debug information can be taken into account. To
make that happen, we have to first understand which symbols belong to
kernel modules and which to the kernel proper.
With this change we extend the Kfunc type with the mod_ member capturing
this information, which itself is derived from parsing kallsyms.