Add parser for /proc/modules
Add a parser for /proc/modules style files, ModMap. /proc/modules
contains information about memory ranges of modules. We need such
information in order to understand which module a given kernel address
belongs to. On top of just parsing contents, objects of this type
provide the means for querying the module name that an address belongs
to, if any.

Signed-off-by: Daniel Müller <deso@posteo.net>