Resolve kernel module addresses with DWARF
Add support for looking up and incorporating DWARF information for
symbolization of kernel module addresses. So far we have only supported
using DWARF for addresses belonging to the core kernel. We rely on the
previously introduced /proc/modules map for deciding whether an address
belongs to a kernel module. We then use the depmod file for looking up
the kernel module, optionally taking into account debug links as usual.

Signed-off-by: Daniel Müller <deso@posteo.net>