Resolve debug symbols in containers
For the ELF file itself blazesym does the right thing by reading files like:

```
/proc/2367/map_files/79c8000-20aec000
```

This works great when the ELF files are in a container in a separate mount
namespace. When it comes to looking for split debug info via `debuglink`,
it tries looking in by the path in the root mount namespace, which does not
work, as the files are just not there. With this change we try checking
the process mount namespace first via `/proc/<xxx>/root`.

Signed-off-by: Ivan Babrou <github@ivan.computer>